refactor remote sync handler
